scheduler/requirements.txt
Flask==2.2.2
Flask-APScheduler==1.12.4
psycopg2-binary==2.9.5